ajout setvar & tools_Var dans tools/ // merge avant de finir background
