+++ srcs/evaluator/tools_var.c
@@ -17,16 +17,6 @@ uint8_t		ft_print_env_lst(t_list *lst)
 	return (SUCCESS);
 }
 
-void		cpy_var_list(void *dst, void *src)
-{
-	t_var *d;
-	t_var *s;
-
-	d = dst;
-	s = src;
-	d->ctab = ft_tabdup(s->ctab);
-}
-
 char		**create_tab_var(t_list *lst, int count)
 {
 	char	**tab;
@@ -47,30 +37,6 @@ char		**create_tab_var(t_list *lst, int count)
 	var = lst->data;
 	tab[count] = ft_strjoin(3, var->ctab[0], "=", var->ctab[1]);
 	return (tab);
-}
-
-t_list		*find_var(t_list *lst, char *name)
-{
-	t_var *var;
-
-	while (lst)
-	{
-		var = lst->data;
-		if (!ft_strcmp(var->ctab[0], name))
-			return (lst);
-		lst = lst->next;
-	}
-	return (NULL);
-}
-
-char		*find_var_value(t_list *var, char *name)
-{
-	t_list	*lst;
-
-	lst = find_var(var, name);
-	if (lst)
-		return (((t_var *)(lst->data))->ctab[1]);
-	return (NULL);
 }
 
 void		create_lst_var(t_list **lst, char **tab)
