+++ srcs/tools/setvar_add.c
